% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{print hello world + some text}
\usage{
hello(x, separator)
}
\arguments{
\item{x}{A character vector with one element.}

\item{separator}{a string that denotes the separator}
}
\value{
A character vector.
}
\description{
print hello world + some text
}
\examples{
x <- "alpha"
hello(x, separator = "+")
}
